rename swagger users
