+++ swagger/v1/paths/users/regenerateApiKeyUser.js
