+++ swagger/v1/paths/users/registerUser.js
